requirements.txt
streamlit==1.32.2
tensorflow==2.15.0
numpy==1.26.4
opencv-python==4.9.0.80
matplotlib==3.8.3
seaborn==0.13.2
scikit-learn==1.4.1.post1
xgboost==2.0.3
Pillow==10.2.0
